add improve validation to admin controllers

+++ app/Http/Controllers/AdminPostsController.php
@@ -132,7 +132,9 @@ class AdminPostsController extends Controller
     public function destroy($id)
     {
         $post = Post::findOrFail($id);
-        unlink(public_path() . $post->photo->file);
+        if($post->photo->file){
+            unlink(public_path() . $post->photo->file);
+        } 
         $post->delete();
 
         Session::flash("flash_notification", [
